Imrpove code
